Mark ProxyChangeListener as @UsedByReflection.
A number of WebView applications in the wild use reflection to call the
proxy change broadcast receiver directly as a hack to allow them to set
an application-local proxy setting. The reflection code used searches
for ProxyChangeListener by name, so fails with proguard obfuscation
enabled for the monochrome APK. Mark the relevant classes/methods as
@UsedByReflection to ensure they don't get renamed, because while it's
not good that applications do this, there is no supported API to achieve
the same thing either, and it appears to be a commonly repeated piece of
code.

Bug: 851995
Change-Id: I6174181491f57fd71fece91c98ae8b96f36b425f
Reviewed-on: https://chromium-review.googlesource.com/1112564
Reviewed-by: Paul Jensen <pauljensen@chromium.org>
Commit-Queue: Richard Coles <torne@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#570479}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0f71a543910198ee5b4104ea0066494a3fcba309