Reporting: Don't assume report and policy have same origin
When processing a report, we can't assume that the report's origin
matches the origin that the corresponding policy belongs to.  If the
policy uses include_subdomains, then it might be used for reports for
any subdomain of the policy's origin.  We had a couple of places where
we did assume they were the same, and used a DCHECK to verify this.
That caused segfaults whenever we tried to process a subdomain report.

Bug: 854248
Change-Id: I251d82664c7deee2293144ceed199e084795152c
Reviewed-on: https://chromium-review.googlesource.com/1106304
Reviewed-by: Matt Menke <mmenke@chromium.org>
Commit-Queue: Douglas Creager <dcreager@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#571946}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e13028c51bdd8b4e3a3373a6374fc69b06ef02c4