Enable parsing of DNS queries and writing of DNS responses.
This CL adds utilities to DnsQuery to allow the parsing of a DNS
query in the wire format if it is a valid query with a single question.
A new constructor of DnsResponder is also added to generate DNS
responses in the wire format with multiple resource record as answers.
Name compression in multiple answers are not yet implemented in this CL.

Combined with the existing utilities in //net/dns/, we now have a full
set of capabilities to parse and generate DNS/mDNS queries and response.

Bug: 878465
Change-Id: Ifeae0d17243df037af7a3ea0bffb1079018ea714
Reviewed-on: https://chromium-review.googlesource.com/1203320
Commit-Queue: Qingsi Wang <qingsi@google.com>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: Eric Orth <ericorth@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#594877}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ecd69251f0ccbefee3d3f128054f8f817324383c