Fix object-lifetime comments in net/socket/udp_socket_*.h
Fixes the comments on RecvFrom and SendTo in udp_socket_posix.h and
udp_socket_win.h to more accurately reflect who is responsible for
keeping some of the function parameters alive.

Namely, the sockets always take a reference to the provided IOBuffer
in the case of asynch IO. And the SendTo function also creates a copy
of the provided IPAddress if necessary, while RecvFrom take a raw
pointer to the IPAddress and the caller is in charge of keeping it
alive.

Change-Id: I080557b8c69c9035eace94a0359ec5ce09c4cf0d
Reviewed-on: https://chromium-review.googlesource.com/c/1290089
Reviewed-by: Matt Menke <mmenke@chromium.org>
Commit-Queue: Matthew Denton <mpdenton@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#601286}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 111df5adb031caba145484aaddc5a561c52f0e56