Servicified custom DNS configuration
Mojo API to override system DNS config for built-in resolver
resolutions.  This is the servicification of the API added in CL
1259902.

When a non-nullopt overrides struct is specified in CreateHostResolver
calls, a separate internal net::HostResolver is created and used with
the overrides. A better solution would be to use the normal shared
net::HostResolver, eg by setting per-request overrides, but that is
unnecessary and out of scope for the current needs where this is only
expected to be used by a special case (DnsProbeService) that doesn't
care about shared stuff like caching.

TBR=caseq@chromium.org

Bug: 846423
Cq-Include-Trybots: luci.chromium.try:linux_mojo
Change-Id: I1b49da83d166f80b503bd6955328f3492d553a5e
Reviewed-on: https://chromium-review.googlesource.com/c/1269076
Commit-Queue: Eric Orth <ericorth@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#602367}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e366fe60c014a059ccbef6d6db48fe47c33de20a